infra: Make docker folder read only, remove unused tests.
* Remove container tests.
* Rename buildspec.yml -> buildspec-container-pr.yml.
* Rename buildspec-toolkit.yml -> buildspec.yml.
* Rename test-toolkit/ -> test/.
* Refactor tests.
