+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
-application_import_names = image_utils, integration, sagemaker_tensorflow_container, test, test-toolkit, timeout, utils
+application_import_names = image_utils, integration, sagemaker_tensorflow_container, test, timeout, utils
 import-order-style = google
