+++ test/integration/__init__.py
@@ -1,4 +1,4 @@
-#  Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#  Copyright 2018-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License").
 #  You may not use this file except in compliance with the License.
@@ -20,30 +20,3 @@ logging.getLogger('botocore').setLevel(logging.INFO)
 
 RESOURCE_PATH = os.path.join(os.path.dirname(__file__), '..', 'resources')
 DEFAULT_TIMEOUT = 120
-
-# these regions have some p2 and p3 instances, but not enough for automated testing
-NO_P2_REGIONS = [
-    'ca-central-1',
-    'eu-central-1',
-    'eu-west-2',
-    'us-west-1',
-    'eu-west-3',
-    'eu-north-1',
-    'sa-east-1',
-    'ap-east-1',
-    'me-south-1'
-]
-NO_P3_REGIONS = [
-    'ap-southeast-1',
-    'ap-southeast-2',
-    'ap-south-1',
-    'ca-central-1',
-    'eu-central-1',
-    'eu-west-2',
-    'us-west-1'
-    'eu-west-3',
-    'eu-north-1',
-    'sa-east-1',
-    'ap-east-1',
-    'me-south-1'
-]
