--- test/resources/test_py_version/entry.py
@@ -1,22 +1,0 @@
-# Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License"). You
-# may not use this file except in compliance with the License. A copy of
-# the License is located at
-#
-#     http://aws.amazon.com/apache2.0/
-#
-# or in the "license" file accompanying this file. This file is
-# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
-# ANY KIND, either express or implied. See the License for the specific
-# language governing permissions and limitations under the License.
-from __future__ import absolute_import
-
-import os
-import sys
-
-
-py_version = '%s.%s' % (sys.version_info.major, sys.version_info.minor)
-
-with open(os.path.join(os.environ['SM_OUTPUT_DIR'], 'py_version'), 'a') as f:
-    f.write(py_version)
