+++ tox.ini
@@ -22,7 +22,6 @@ exclude =
     .tox
     tests/data/
     test/resources
-    test-toolkit/resources
     venv/
     sagemaker-tensorflow-extensions
     benchmarks/
